% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-funcs.R
\name{read_pkmx}
\alias{read_pkmx}
\title{Read psykinematix file}
\usage{
read_pkmx(path, type = NULL, ...)
}
\arguments{
\item{path}{string. path to psykinematix file}

\item{type}{string. override extension detection}

\item{...}{other arguments to the read functions}
}
\value{
depending on read function called, data or string vector
}
\description{
Convenience function, will automatically detect
file type based on file extension and use
the correct reading function
}
